move SchemaCommands into webapi
